client-cli/py/docutil/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docutil"
version = "0.1.0"
description = "DOCX document analysis and formatting utilities"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "EventAI Project"}
]
dependencies = [
    "python-docx>=1.0.0",
    "click>=8.0.0",
]

[project.scripts]
docutil = "docutil.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["docutil*"]
